Implemented basic rendering

+++ common/src/proto/io.rs
@@ -346,14 +346,14 @@ array!(4096);
 impl ProtoType for crate::types::Vector {
     const CHECKSUM: u128 = {
         #[derive(codegen::Gen)]
-        struct _Vector([f32; 3]);
+        struct _Vector([f64; 3]);
         _Vector::CHECKSUM
     };
 }
 
 impl BinRead for crate::types::Vector {
     fn read(buf: &mut &[u8]) -> Result<Self> {
-        let inner = <[f32; 3]>::read(buf)?;
+        let inner = <[f64; 3]>::read(buf)?;
         Ok(Self::from_column_slice(&inner[..]))
     }
 }
@@ -361,7 +361,7 @@ impl BinRead for crate::types::Vector {
 impl BinWrite for crate::types::Vector {
     fn write(&self, vec: &mut Vec<u8>) {
         let slice = self.as_slice();
-        let array: [f32; 3] = slice.try_into().expect("Vector has exactly 3 elements");
+        let array: [f64; 3] = slice.try_into().expect("Vector has exactly 3 elements");
         array.write(vec);
     }
 }
@@ -369,14 +369,14 @@ impl BinWrite for crate::types::Vector {
 impl ProtoType for crate::types::Matrix {
     const CHECKSUM: u128 = {
         #[derive(codegen::Gen)]
-        struct _Matrix([f32; 16]);
+        struct _Matrix([f64; 16]);
         _Matrix::CHECKSUM
     };
 }
 
 impl BinRead for crate::types::Matrix {
     fn read(buf: &mut &[u8]) -> Result<Self> {
-        let inner = <[f32; 16]>::read(buf)?;
+        let inner = <[f64; 16]>::read(buf)?;
         Ok(Self::from_column_slice(&inner[..]))
     }
 }
@@ -384,7 +384,7 @@ impl BinRead for crate::types::Matrix {
 impl BinWrite for crate::types::Matrix {
     fn write(&self, vec: &mut Vec<u8>) {
         let slice = self.as_slice();
-        let array: [f32; 16] = slice.try_into().expect("Matrix has exactly 16 elements");
+        let array: [f64; 16] = slice.try_into().expect("Matrix has exactly 16 elements");
         array.write(vec);
     }
 }
